@@ -6,20 +6,23 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack-dev-server --hot --inline --progress --colors",
-    "build": "webpack --progress --colors"
+    "start1": "webpack-dev-server --hot --progress --colors",
+    "start3": "webpack --progress --colors",
+    "plane": "webpack",
+    "dev": "webpack-dev-server --devtool eval"
   },
   "keywords": [
     "EncryptIndexedDBEntityManager"
   ],
   "author": "RyunosinFX",
   "license": "MIT",
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
     "css-loader": "^0.25.0",
-    "eslint": "^3.8.1",
+    "eslint": "^3.19.0",
     "eslint-loader": "^1.6.0",
     "file-loader": "^0.9.0",
+    "snabbdom": "^0.7.0",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.3",
     "webpack-dev-server": "^1.16.2"
